Factor out MediaUIShareLink
